améliore la gestion des erreurs + récupération des objets depuis le serveur

+++ src/app/user-list/user-list.component.ts
@@ -1,7 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {User} from "../shared/data.model";
 import {DataService} from "../shared/data.service";
-import {FeedbackService} from "../shared/feedback/feedback.service";
 
 @Component({
   selector: 'app-user-list',
@@ -12,22 +11,24 @@ export class UserListComponent implements OnInit {
 
   displayedColumns: string[] = ['id', 'name', 'role', 'action'];
 
-  constructor(public dataService: DataService, private feedbackService: FeedbackService) {
+  constructor(public dataService: DataService) {
   }
 
-  users : User[];
+  users: User[];
 
   ngOnInit(): void {
     this.getUsers();
   }
 
-  getUsers(){
+  getUsers() {
     this.dataService.getUsers().subscribe(users => this.users = users);
   }
 
 
   delete(id: number) {
-    this.dataService.deleteUser(id).subscribe();
+    if (confirm('Are you sure?')) {
+      this.dataService.deleteUser(id).subscribe();
+    }
   }
 
 }
